Introduce a SqliteConnection wrapper.
It allows us not to share connections between threads, and will do for
the time being.
It's quite unsafe, as TLS is used to store the connection, but doesn't
mark the connections as released.
This is mostly commited as such to move forward, we'll have to rewrite
the databases interactions in a near future
Now I'll be in the corner, crying.