IMediaLibrary: Improve onReload* callbacks
Don't invoke the callback with an empty entry point and a global success
state.
Instead, let the discoverer invoke the callback with the actual entry
point being reloaded, and its associated success state
The deleteFolder test was removed because the RemoveDirectory test
already covers folder deletion tests, but also because the folder was
directly removed from database, which makes the reload operation not
reload anything, so no callback was ever sent, and waitForReload was
timing out.