Playlist: Fix & simplify insert/delete triggers
When inserting, we just need to ensure the trigger won't modify the
newly inserted row. When deleting, we don't care about the old row since
it's already gone when the trigger fires.
What we always care about is the position of the modified items relative
to the inserted/deleted row. The media ID doesn't matter since a single
media ID might be inserted multiple times in a single playlist