MetadataParser: Copy all thumbnails for now
Until libvlc allows us to specifically request the attachment, we can't
discriminate between an album thumbnail (ie. cover.jpg and the likes)
and an attachment cover. We could only use the medialibrary album cover
extraction, but this would mean we're dropping attached thumbnails.

This commit is meant to be reverted ASAP

fix #51