MetadataParser: Rework thumbnail handling for audio content
We now favor album thumbnails for album & media. If an album cover is
available, it will be assigned to the album and all media, even if they
have an embedded cover.
We also propagate embedded covers to all media & the album if they
didn't have embedded artwork or a cover file available