utils: Directory: Remove superfluous exception handling
We are manipulating a Directory object that was created based on a local
path, so the exception that can be thrown from toLocalPath will never be
thrown here.
If url::decode throws for an mrl that was constructed just before (ie.
not stored in DB) we're in trouble, and this indicates a code problem
rather than some invalid user input