CI: Disable win32 tests
The dwarf exception handling doesn't work under wine in docker, because
it requires process introspection, which, when run on wine implies
using ptrace.
This causes EnumProcessModules to fail in libunwind, so we're not able
to unwind the stack properly, causing all exceptions to go unhandled.
Until we can either run the docker image with SYS_PTRACE capability, or
wine uses a different way of running EnumProcessModules when the target
process is the running one, I'm not sure we can run the win32 tests in
the CI.
win64 isn't affected, as it uses SEH exception model