IFileSystemFactory: Simplify callback prototypes
No need to take the new mountpoint as a parameter since this information
isn't valuable to the media library. What matters is the device
presence.
In addition, we now return a boolean from this callback to ensure any
potential device lister can know if a device was seen for the
first time or not
refs #216