Parser: Fix linking tasks retry handling
The sql request that fetches the tasks allow a retry count to less than
or equal to the max number of retries, so in order for all tasks
behavior to be consistent, we have to apply the same comparison when
handling linking tasks at runtime.
This is especially important since we reduced the number of retries to
1. The current code meant that if linking failed, we would never try
again.
This is likely to improve our sporadic functionnal test failures