MediaLibrary: Add thumbnail generation API
Add an API to request asynchronous thumbnail generation from a media id.

The existing API (IMedia::requestThumbnail()) needs a Media instance.
If the client only knows the media id, it must execute a SQL query to
retrieve a Media, so for an UI client, this defeated the purpose of an
asynchronous API.

This new API resolves the media id from the worker thread.