MetadataParser: Simplify playlist import
We now always try to fetch a media for each MRL we try to insert
into a playlist. If there's none, we immediatly create an external
media and spawn a link task for that external media.
This removes the need for a specific discovery configuration when
crawling path contained in a playlist.
refs #303