tests: samples: Simplify locking
We used to need a lock to be held before a discovery was scheduled and
at least until after the parsing was complete.
This is not the case anymore since the discoverer thread only emits a
discovery completed event upon full completion, rather than each time a
specific entry point gets discovered so we don't need to update various
states from multiple callback in order to wait for all folders to be
discovered and parsing to be completed after that.
This removes a potential lock inversion and simplifies the tests a bit