Album: Split empty album deletion in a dedicated trigger
Since an album track is denoted by the media.subtype column, we will
need to monitor this field as well to update the nb_tracks field, and
can't assume that an album track removal will lead to a media deletion.
The subtype is more likely to be updated.

It's also clearer to have a dedicated trigger for a dedicated task

refs #372