CI: Disable 'needs: ' rule on benchmark jobs
This way, the benchmark jobs won't start if the checks fail, or if the
debian tests fail.
If specifying `needs: []` the job doesn't start when it shouldn't, but
when it starts, it uses all artifacts from all previous test jobs, which
causes warning at best, or tries to use the windows artifacts.