CI: Remove automated benchmark.
I can't find a proper way of having a nightly benchmark that doesn't
require to regenerate & discover samples during each run, or that don't
require manual intervention each time the database model is changed.
Until we can have some actions when a merge request with a specific
label is merged (in which case we could regenerate the model when the
new model gets merged to master) I'm not sure there's a good way of
doing this.
In addition, when the benchmark is run as part of a MR, the results are
too dependent on the current CI machine load.
Refs #389