tests: Fix data race on vptr
When tearing down MediaLibraryWithDiscoverer, we end up stopping the
parser from the media library destructor (not the
MediaLibraryWithDiscoverer one) which causes the vptr to be updated to
the MediaLibrary one. This will cause the regular getParser overload to
be used if the discoverer thread asks for a parser when stopping.
By explicitly stopping the discoverer from the test class, we ensure the
proper vptr gets used until the discoverer thread joins