Ensure CheckDbModel always acquires a read context
This is likely suboptimal at the moment since it will acquire many
individual read context while the main calls to those functions are to
check all tables models, and it should do so with a single read context
from that common call site.
However the individual methods need to acquire a context for their own
unit tests as well and the tests have no business acquiring a context if
they don't execute a manual request.