Media: Change sort and join requests prototypes
To allow selecting the default `SortingCriteria` which was annoying to
do beforehand because of the `const` nature of the `QueryParameters`.

This will be useful to set `Album` sorting as a default for `Artist`
tracks listing.

(cherry picked from commit d142bb042e06f26387d9f5f2f5d5a8363b5cf4e9)