Media: Notify external media creation
Users actually expect the "New Media" notification to trigger at each
media addition instead of just the internal ones.

This was reported by the Android team as they are building a remote
access service. Notifying external media creation this way allow them to
update their UI without having to worry if the media is internal or
external. It's simpler that way.